Fix page section and title structure for clusters
